+++ src/gameobjects/DynamicCircle.java
@@ -18,7 +18,7 @@ public class DynamicCircle extends Circle {
      * @param friction The friction of the body
      * @param image The image representing the body in the visual realm
      */
-    public DynamicCircle(int ID, World world, Vec2 pos, float friction, Image image) {
+    public DynamicCircle(long ID, World world, Vec2 pos, float friction, Image image) {
         super(ID, world, pos, friction, image);
         body.setUserData(this);
         makeDynamic();
@@ -31,7 +31,7 @@ public class DynamicCircle extends Circle {
      * @param friction The friction of its body
      * @param color The color representing the body in the visual realm
      */
-    public DynamicCircle(int ID, World world, Vec2 pos, float friction, Color color, double radious) {
+    public DynamicCircle(long ID, World world, Vec2 pos, float friction, Color color, double radious) {
         super(ID, world, pos, friction, color, radious);
         body.setUserData(this);
         makeDynamic();
