+++ src/gameobjects/DynamicSquare.java
@@ -18,7 +18,7 @@ public class DynamicSquare extends Square {
      * @param friction The friction of the body
      * @param image The image representing the body in the visual realm
      */
-    public DynamicSquare(int ID, World world, Vec2 pos, float friction, Image image) {
+    public DynamicSquare(long ID, World world, Vec2 pos, float friction, Image image) {
         super(ID, world, pos, friction, image);
         body.setUserData(this);
         makeDynamic();
@@ -33,7 +33,7 @@ public class DynamicSquare extends Square {
      * @param width The width of the body in meters
      * @param height The height of the body in meters
      */
-    public DynamicSquare(int ID, World world, Vec2 pos, float friction, Color color, double width, double height) {
+    public DynamicSquare(long ID, World world, Vec2 pos, float friction, Color color, double width, double height) {
         super(ID, world, pos, friction, color, width, height);
         body.setUserData(this);
         makeDynamic();
